requirements.txt
grequests
wheel
mysql-connector-python
Django
mysqlclient
djangorestframework
